# Generalize special casing of pluralized acronyms