--- a/botocore/__init__.py
+++ b/botocore/__init__.py
@@ -33,7 +33,7 @@ _number_cap_regex = re.compile('([a-z])([0-9]+)')
 _end_cap_regex = re.compile('([a-z0-9])([A-Z])')
 # The regex below handles the special case where some acryonym
 # name is pluralized, e.g GatewayARNs, ListWebACLs, SomeCNAMEs.
-_special_case_transform = re.compile('[A-Z][A-Z][A-Z]+s$')
+_special_case_transform = re.compile('[A-Z]{3,}s$')
 # Prepopulate the cache with special cases that don't match
 # our regular transformation.
 _xform_cache = {