Given that this is in the critical path for making API calls,
I did some benchmarking to ensure that this doesn't cause any
serious performance issues.  Short answer is that this seems
acceptable.

I timed how long it took to xform every single operation and
top level member name against this branch and the develop
branch.

Given 3034 xform's, I'm seeing:

```
dev branch   - 0.0930128 seconds averaged over 10 runs of 3034
fix-arn-name - 0.0944845 seconds averaged over 10 runs of 3034
```

Or in terms of ops/sec:

```
before - 32619.2 ops/sec
after  - 32111.1 ops/sec
```

The reason I think this should be generalized is that some
services use this pattern for particular types of resources
(mostly ACLs and ARNs), and we'll need to possibly update
our special cases every time we pull in a service update.

With this change, we won't need to do this.  I will need
to make a corresponding PR to the AWS CLI to put in
backward compat shims so we can still support both.

Using [this script](https://gist.github.com/jamesls/0d1005415d30d2f9f44a),
the delta between the develop branch and this PR is:

```
Model Name                                          xform name

-OpenIdConnectProviderARNs                          open_id_connect_provider_ar_ns
+OpenIdConnectProviderARNs                          open_id_connect_provider_arns

-TapeARNs                                           tape_ar_ns
+TapeARNs                                           tape_arns

-VTLDeviceARNs                                      vtl_device_ar_ns
+VTLDeviceARNs                                      vtl_device_arns

-VolumeARNs                                         volume_ar_ns
+VolumeARNs                                         volume_arns

```


cc @kyleknap @mtdowling @rayluo @JordonPhillips